# Improve CryptoConfig.CreateFromName performance (from 10% to x20)