I was working on making it easier for CoreFX team members to use ASP.NET Perf Lab to run performance tests under high load.

Then I saw #39600 which was a perfect candidate for sample for the docs: we had some nasty locks that were making the perf bad. So I created sample docs based on this example: https://github.com/aspnet/Benchmarks/pull/1284 

After taking a brief look at the existing code, I could not resist rewriting it and make it even faster than @VladimirKhvostov did in #39600 


ASP.NET Perf lab test result (described in https://github.com/aspnet/Benchmarks/pull/1284):

Before:

```log
RequestsPerSecond:           43,330
Max CPU (%):                 95
WorkingSet (MB):             166
Avg. Latency (ms):           5.91
```

<img width="960" alt="flamegraph_filtered" src="https://user-images.githubusercontent.com/6011991/62784141-16c3e400-babe-11e9-8a91-6ee02a6ef550.png">

After:

```log
RequestsPerSecond:           120,801
Max CPU (%):                 90
WorkingSet (MB):             166
Avg. Latency (ms):           2.27
```

<img width="963" alt="flamegraph_filtered_after" src="https://user-images.githubusercontent.com/6011991/62784153-1a576b00-babe-11e9-82f7-13a55e9d70aa.png">


Some [microbenchmark](https://github.com/dotnet/performance/blob/master/src/benchmarks/micro/corefx/System.Security.Cryptography/Perf.CryptoConfig.cs) results:

|         Method |     Toolchain |      name |       Mean | Ratio | Allocated Memory/Op |
|--------------- |-------------- |---------- |-----------:|------:|--------------------:|
| CreateFromName |         net48 |       RSA |   765.4 ns |  1.00 |               802 B |
| CreateFromName | netcoreapp3.0 |       RSA | 3,681.8 ns |  4.81 |               688 B |
| CreateFromName |       this PR |       RSA |   278.1 ns |  0.36 |               192 B |
|                |               |           |            |       |                     |
| CreateFromName |         net48 |    SHA512 |   605.7 ns |  1.00 |              1316 B |
| CreateFromName | netcoreapp3.0 |    SHA512 |   770.7 ns |  1.27 |               424 B |
| CreateFromName |       this PR |    SHA512 |   545.3 ns |  0.90 |               128 B |
|                |               |           |            |       |                     |
| CreateFromName |         net48 | X509Chain | 4,188.9 ns |  1.00 |               682 B |
| CreateFromName | netcoreapp3.0 | X509Chain | 3,269.7 ns |  0.78 |               464 B |
| CreateFromName |       this PR | X509Chain |   228.1 ns |  0.05 |                88 B |

@danmosemsft this was one of the cases where Full .NET Framework was faster than Core. I made sure that Core is faster now.

The key changes:

* remove the locks, use ConcurrentDictionary instead
* `Type.GetType(string name)` is very expensive. Every type is now loaded once and then saved to the dictionary. This was the biggest improvement
* one dictionary for name to name mapping, one for name to type mapping (previously there was one for string to object)
* Fast path for creating config without providing any extra arguments