@stephentoub 2019-08-09
@adamsitnik, how do your changes relate to https://github.com/dotnet/corefx/pull/39600?  I've not reviewed yours, but from the description it seems duplicative.