@GrabYourPitchforks 2019-08-09
Are any real-world E2E scenarios improved by this change? I'd be surprised to learn that this function is the bottleneck in any legitimate use of cryptography.