@VladimirKhvostov 2019-08-09
@GrabYourPitchforks this method is used in the JWT token validation code path, which caused several LSIs in other DevOps. See #39600 for details.