@GrabYourPitchforks 2019-08-09
@VladimirKhvostov I saw that, but it seems the conclusion from that thread was to change the JWT verification implementation to call a different method. To an extent this matches recommended practice: the CryptoConfig class is really intended for use only when you need agility. JWT verification shouldn't be bouncing through this class because the JWT validation library should be hardcoding the list of supported algorithms; it doesn't require the type of agility the CryptoConfig class is intended to provide.