@krwq 2019-08-09
@adamsitnik thanks! I think next steps would be to not call CryptoConfig.CreateFromName at all and just use algorithm directly. That would avoid any lookup at all