@krwq 2019-08-09
I have some security concerns about this fix (and also original PR) - will need to discuss it with @bartonjs since he's more familiar with how this is being usage. I'll take this offline