@bartonjs 2019-08-13
#39600 has been accepted.  If there're any more gain to be gotten from this PR it would need to be re-evaluated in terms of the new baseline.  Since it'd be a different change with different benchmark justification, I'm closing this PR.